requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-chroma
langgraph
langchain-core
httpx>=0.25.2
python-dotenv
pydantic
typing-extensions
pypdf
tiktoken